[upspinserver] added field for keyserver in config
With the decomissioning of the upspin.io servers,
it becomes necessary to run a dedicated keyserver
and point to it in the configuration of various
upspin components.

While this is possible for the upspin-client and
their configuration file (usually in ~/upspin/config),
the configuration file (serverconfig.json) for
various storage servers does not have an option to
refer to a keyserver and therefore always use the
hardcoded 'keys.upspin.io:443' for key lookups.

This commit adds a field 'KeyServer' to the server
configuration, as upspin.NetAddr, and when set,
will be used in upspinserver implementations.

Note:  This is commit is an intermediate step towards
two additional, complementary goals:

1.) implement a local inproc-keyserver, embedded into
    the services, reading upspin.User from local files.

2.) add DNS-based lookup of a keyserver for a domain